+++ src/grscheller/datastructures/__init__.py
@@ -18,7 +18,7 @@
    endeavor to be Pythonic.
 """
 
-__version__ = "0.10.9.1"
+__version__ = "0.10.10.0"
 __author__ = "Geoffrey R. Scheller"
 __copyright__ = "Copyright (c) 2023 Geoffrey R. Scheller"
 __license__ = "Appache License 2.0"